+++ jest.config.js
@@ -5,6 +5,7 @@ const tsJestTransformCfg = createDefaultPreset().transform;
 /** @type {import("jest").Config} **/
 module.exports = {
   testEnvironment: "node",
+  testPathIgnorePatterns: ["/node_modules/", "<rootDir>/dist/"],
   transform: {
     ...tsJestTransformCfg,
   },